laser_rabi: rename to laser_flop, update __init__files
