+++ ionics_fits/__init__.py
@@ -1,4 +1,4 @@
-from .common import ModelParameter, Model, Fitter
+from .common import Fitter, Model, ModelParameter
 from .normal import NormalFitter
 from . import models
 from . import utils
